pincode-fraud-detection/requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone-client
numpy